refactors slack flow, changes helpers to util
